Split up sdcard G-codes
