+++ Marlin/src/gcode/sdcard/M28_M29.cpp
@@ -1,0 +1,73 @@
+/**
+ * Marlin 3D Printer Firmware
+ * Copyright (C) 2019 MarlinFirmware [https://github.com/MarlinFirmware/Marlin]
+ *
+ * Based on Sprinter and grbl.
+ * Copyright (C) 2011 Camiel Gubbels / Erik van der Zalm
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+#include "../../inc/MarlinConfig.h"
+
+#if ENABLED(SDSUPPORT)
+
+#include "../gcode.h"
+#include "../../sd/cardreader.h"
+
+/**
+ * M28: Start SD Write
+ */
+void GcodeSuite::M28() {
+
+  #if ENABLED(BINARY_FILE_TRANSFER)
+
+    bool binary_mode = false;
+    char *p = parser.string_arg;
+    if (p[0] == 'B' && NUMERIC(p[1])) {
+      binary_mode = p[1] > '0';
+      p += 2;
+      while (*p == ' ') ++p;
+    }
+
+    // Binary transfer mode
+    if ((card.flag.binary_mode = binary_mode)) {
+      SERIAL_ECHO_START();
+      SERIAL_ECHO(" preparing to receive: ");
+      SERIAL_ECHOLN(p);
+      card.openFile(p, false);
+      #if NUM_SERIAL > 1
+        card.transfer_port_index = command_queue_port[cmd_queue_index_r];
+      #endif
+    }
+    else
+      card.openFile(p, false);
+
+  #else
+
+    card.openFile(parser.string_arg, false);
+
+  #endif
+}
+
+/**
+ * M29: Stop SD Write
+ * (Processed in write-to-file routine)
+ */
+void GcodeSuite::M29() {
+  card.flag.saving = false;
+}
+
+#endif // SDSUPPORT
